final completion of search use case without testing
