--- src/users/service/search_recipes/API_Interface/APICaller.java
@@ -1,5 +1,0 @@
-package users.service.search_recipes.API_Interface;
-
-public interface APICaller {
-    public OutputData execute(String[] urlComponents);
-}
