--- src/users/service/search_recipes/API_Interface/InputDataURLConverter.java
@@ -1,7 +1,0 @@
-package users.service.search_recipes.API_Interface;
-
-import users.service.search_recipes.SearchInputData;
-
-public interface InputDataURLConverter {
-    public String convertToURL(SearchInputData searchInputData);
-}
