--- src/users/service/search_recipes/JsonOutputDataConverter.java
@@ -1,2 +1,0 @@
-package users.service.search_recipes;public interface JsonOutputDataConverter {
-}
