--- src/users/service/search_recipes/SearchInputBoundary.java
@@ -1,2 +1,0 @@
-package users.service.search_recipes;public interface SearchInputBoundary {
-}
