+++ src/users/service/search_recipes/application_business_rules/API_Interface/APICaller.java
@@ -1,0 +1,5 @@
+package users.service.search_recipes.application_business_rules.API_Interface;
+
+public interface APICaller {
+    public String execute(String pullRequestURL);
+}
